refactor(frontend): move the serialized Directus fetch queue to lib/, shared beyond PlaylistEditor
