+++ packages/frontend/src/Applications/PlaylistEditor/directusVolume.ts
@@ -6,7 +6,7 @@ import {
 	PRESIDENTIAL_FLIGHTS,
 } from "../FlightTracker/notableFlights";
 import { BROADCAST_STATIONS } from "../radio-core/stationGrouping";
-import { directusGet } from "./directusQueue";
+import { directusGet } from "../../lib/directusQueue";
 
 export const MEDIA_FILE_TYPES = {
 	tv: "tv-channel",
